requirements.txt
farm-haystack==1.25.4
torch==2.2.2